Tagged dest directory detection
Add detection of existing directory with the same format but tagged,
for example, if we are trying to move a picture to

pictures/2020/2020_02_03/

and a directory already exist with the following format:

pictures/2020/2020_02_03_party_at_mikes/

The picture will be copied to such directory instead. With this
we support directores which have been renamed to be more descriptive.